Create rainfallData.java
Class assignment that simulates a rainfall data collector program which allows the user to enter in values for rainfall data into an array that stores 52 values. I chose to use a random number generator because i know that no real person was going to use this program or enter in 52 values. 
1)Random number generator stores 52 values into rainFall[] array
2)Outputs the average of the numbers that are in the array, the largest number in the array, the smallest number in the array, all of the numbers that are in the array. 
This program uses methods and arrays.