Create searchingArrays.java
This program starts out with an unsorted array list. 
Takes in user integer input. 
Outputs if the user input matches any of the numbers in the array. 
Outputs the number of times the users input occurs within the array. 
Outputs the position or positions that the users input occurs within the array. 

This program uses arrays, methods boolean if statements, input and output.