Create TimeSheet.java
A very simple program i did for an online discussion about making and using a Java Class. The program uses the created workWeek class, assigns values to the name,hourlywage and hours within the class and returns the name and hourlywage * hoursworked to output the final value which is the net pay.