Create LeapYear.java
This is a leap year program which tells the user if the users input is a leap year or not. This programs uses input,output and boolean if statements. I didn't get full credit for this assignment because i was told that program isn't 100% correct is determining the leap year. Nearly correct but not 100%.